+++ component/site/libraries/jeventcal.php
@@ -753,7 +753,7 @@ class jEventCal {
 
 	function viewDetailLink($year,$month,$day,$sef=true, $Itemid=0){
 		$Itemid	= $Itemid>0?$Itemid:JEVHelper::getItemid($this);
-		$title = JApplication::stringURLSafe($this->title());
+		$title = JApplicationHelper::stringURLSafe($this->title());
 		$link = "index.php?option=".JEV_COM_COMPONENT."&task=".$this->detailTask()."&evid=".$this->id() .'&Itemid='.$Itemid
 		."&year=$year&month=$month&day=$day" ;
 		if (JRequest::getCmd("tmpl","")=="component" && JRequest::getCmd('task', 'selectfunction')!='icalevent.select'  && JRequest::getCmd("option","")!="com_acymailing" && JRequest::getCmd("option","")!="com_jnews" && JRequest::getCmd("jevtask","")!="crawler.listevents"){